Invoice generation: Add a hack to make invoice generation possible after SIMS have been allocated to accounts.
Multiple SIM records with the same ICCID (but not overlapping periods) may
exist, after a SIM has been changed. In this case, the old code would not only
update the "current" record but also STOCK records, causing an exception to be
thrown. This hack only updates the current record. The right fix is documented
in the code for later.