invoice-generate: replace SIM "last invoiced" information updating hack with a call to the sim-change tool.
Previously, a SIM with a scheduled update after the invoice date would lead to
that SIM losing its "last invoiced" information on the date of the scheduled
update, because not all future rows where updated, only the current one. This
could cause the SIM to be invoiced multiple times for the same month. (This
hasn't happened, because we check our invoices by hand for sanity.) Now, in the
same situation, sim-change will call die(), preventing this from happening in
the first place. The full fix is implementing history changing support in
sim-change, but this is a more difficult problem that will be addressed later.